# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "chess api for playing chess online.",
        "title": "Chess API",
        "contact": {},
        "license": {
            "name": "MIT"
        }
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Log into an account using provided username and password. And get an API key.\nUsername can be between 3-20 characters.\nPassword must be at least 3 characters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Log into an account and get an API key.",
                "parameters": [
                    {
                        "description": "Login Account",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.UserCredentials"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.ApiKeyResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid username/password",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorReason"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorReason"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.ApiKeyResponse": {
            "type": "object",
            "properties": {
                "apiKey": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorReason": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "reason"
                }
            }
        },
...
        "server.UserCredentials": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Password123"
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4,
                    "example": "JohnDoe"
                }
            }
        }
    }
}